+++ sampletv/src/main/java/testlibuiza/uiza/com/sampletv/activity/BrowseErrorActivity.java
@@ -11,7 +11,7 @@
  * or implied. See the License for the specific language governing permissions and limitations under
  * the License.
  */
-package testlibuiza.uiza.com.sampletv;
+package testlibuiza.uiza.com.sampletv.activity;
 
 import android.app.Activity;
 import android.app.Fragment;
@@ -23,6 +23,9 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.widget.FrameLayout;
 import android.widget.ProgressBar;
+
+import testlibuiza.uiza.com.sampletv.R;
+import testlibuiza.uiza.com.sampletv.frm.ErrorFragment;
 
 /*
  * BrowseErrorActivity shows how to use ErrorFragment
