+++ sampletv/src/main/java/testlibuiza/uiza/com/sampletv/activity/MainActivity.java
@@ -12,24 +12,21 @@
  * the License.
  */
 
-package testlibuiza.uiza.com.sampletv;
+package testlibuiza.uiza.com.sampletv.activity;
 
+import android.app.Activity;
 import android.os.Bundle;
-import android.support.v4.app.FragmentActivity;
 
-/**
- * Loads {@link PlaybackVideoFragment}.
+import testlibuiza.uiza.com.sampletv.R;
+
+/*
+ * MainActivity class that loads {@link MainFragment}.
  */
-public class PlaybackActivity extends FragmentActivity {
+public class MainActivity extends Activity {
 
     @Override
     public void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        if (savedInstanceState == null) {
-            getSupportFragmentManager()
-                    .beginTransaction()
-                    .replace(android.R.id.content, new PlaybackVideoFragment())
-                    .commit();
-        }
+        setContentView(R.layout.activity_main);
     }
 }
