+++ sampletv/src/main/java/testlibuiza/uiza/com/sampletv/activity/PlaybackActivity.java
@@ -12,19 +12,26 @@
  * the License.
  */
 
-package testlibuiza.uiza.com.sampletv;
+package testlibuiza.uiza.com.sampletv.activity;
 
-import android.app.Activity;
 import android.os.Bundle;
+import android.support.v4.app.FragmentActivity;
 
-/*
- * MainActivity class that loads {@link MainFragment}.
+import testlibuiza.uiza.com.sampletv.frm.PlaybackVideoFragment;
+
+/**
+ * Loads {@link PlaybackVideoFragment}.
  */
-public class MainActivity extends Activity {
+public class PlaybackActivity extends FragmentActivity {
 
     @Override
     public void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_main);
+        if (savedInstanceState == null) {
+            getSupportFragmentManager()
+                    .beginTransaction()
+                    .replace(android.R.id.content, new PlaybackVideoFragment())
+                    .commit();
+        }
     }
 }
