+++ sampletv/src/main/java/testlibuiza/uiza/com/sampletv/model/MovieList.java
@@ -12,7 +12,7 @@
  * the License.
  */
 
-package testlibuiza.uiza.com.sampletv;
+package testlibuiza.uiza.com.sampletv.model;
 
 import java.util.ArrayList;
 import java.util.List;
