+++ sampletv/src/main/java/testlibuiza/uiza/com/sampletv/presenter/DetailsDescriptionPresenter.java
@@ -12,25 +12,22 @@
  * the License.
  */
 
-package testlibuiza.uiza.com.sampletv;
+package testlibuiza.uiza.com.sampletv.presenter;
 
-import android.app.Activity;
-import android.os.Bundle;
+import android.support.v17.leanback.widget.AbstractDetailsDescriptionPresenter;
 
-/*
- * Details activity class that loads LeanbackDetailsFragment class
- */
-public class DetailsActivity extends Activity {
-    public static final String SHARED_ELEMENT_NAME = "hero";
-    public static final String MOVIE = "Movie";
+import testlibuiza.uiza.com.sampletv.model.Movie;
 
-    /**
-     * Called when the activity is first created.
-     */
+public class DetailsDescriptionPresenter extends AbstractDetailsDescriptionPresenter {
+
     @Override
-    public void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_details);
-    }
+    protected void onBindDescription(ViewHolder viewHolder, Object item) {
+        Movie movie = (Movie) item;
 
+        if (movie != null) {
+            viewHolder.getTitle().setText(movie.getTitle());
+            viewHolder.getSubtitle().setText(movie.getStudio());
+            viewHolder.getBody().setText(movie.getDescription());
+        }
+    }
 }
