+++ sampletv/src/main/res/layout/activity_details.xml
@@ -2,8 +2,8 @@
 <fragment xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/details_fragment"
-    android:name="testlibuiza.uiza.com.sampletv.VideoDetailsFragment"
+    android:name="testlibuiza.uiza.com.sampletv.frm.VideoDetailsFragment"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".DetailsActivity"
+    tools:context=".activity.DetailsActivity"
     tools:deviceIds="tv" />
