+++ sampletv/src/main/res/layout/activity_main.xml
@@ -2,9 +2,9 @@
 <fragment xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/main_browse_fragment"
-    android:name="testlibuiza.uiza.com.sampletv.MainFragment"
+    android:name="testlibuiza.uiza.com.sampletv.frm.MainFragment"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainActivity"
+    tools:context=".activity.MainActivity"
     tools:deviceIds="tv"
     tools:ignore="MergeRootFrame" />
